+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 edbee.app:
 
+- add #63, Add duplicate file to the sidebar
 - add #90, Add a config setting to to enable/disable the rendering of whitespaces
 - add #89, Implement smarter backspace for the start of the line
 - add #87, Add setting auto-reveal to keep the file tree-view in sync with active tab